requirements.txt
nltk>=3.1.2
bllipparser>=2016.9.11
flask>=0.12
python-dateutil>=2.6.0